@@ -1,13 +1,13 @@
 module github.com/saichler/servicepoints/go
 
-go 1.23.2
+go 1.24.1
 
 require (
-	github.com/saichler/layer8/go v0.0.0-20250313000847-63d867bf767e
+	github.com/saichler/layer8/go v0.0.0-20250320154019-38f2e89d5827
 	github.com/saichler/reflect/go v0.0.0-20250317153135-a0a75946f793
 	github.com/saichler/serializer/go v0.0.0-20250312171751-f16ab3ede337
 	github.com/saichler/shared/go v0.0.0-20250320123148-50a37cb685f5
-	github.com/saichler/types/go v0.0.0-20250320150223-6b5f73417c34
+	github.com/saichler/types/go v0.0.0-20250320153121-ccf64b8985a1
 	google.golang.org/protobuf v1.36.5
 )
 
