@@ -9,9 +9,10 @@ import (
 	"github.com/saichler/types/go/types"
 )
 
-func CreateNotificationSet(t types.NotificationType, typeName, source string, changeCount int, sequence uint32) *types.NotificationSet {
+func CreateNotificationSet(t types.NotificationType, multicast, protoType, source string, changeCount int, sequence uint32) *types.NotificationSet {
 	notificationSet := &types.NotificationSet{}
-	notificationSet.TypeName = typeName
+	notificationSet.MulticastGroup = multicast
+	notificationSet.ProtoType = protoType
 	notificationSet.Source = source
 	notificationSet.Type = t
 	notificationSet.NotificationList = make([]*types.Notification, changeCount)
@@ -21,11 +22,11 @@ func CreateNotificationSet(t types.NotificationType, typeName, source string, ch
 
 func (this *Cache) createNotificationSet(t types.NotificationType, changeCount int) *types.NotificationSet {
 	defer func() { this.sequence++ }()
-	return CreateNotificationSet(t, this.typeName, this.source, changeCount, this.sequence)
+	return CreateNotificationSet(t, this.multicastGroup, this.protoType, this.source, changeCount, this.sequence)
 }
 
-func CreateAddNotification(any interface{}, typeName, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
-	notificationSet := CreateNotificationSet(types.NotificationType_Add, typeName, source, changeCount, sequence)
+func CreateAddNotification(any interface{}, multicastGroup, protoType, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
+	notificationSet := CreateNotificationSet(types.NotificationType_Add, multicastGroup, protoType, source, changeCount, sequence)
 	obj := object.NewEncode([]byte{}, 0)
 	err := obj.Add(any)
 	if err != nil {
@@ -39,11 +40,11 @@ func CreateAddNotification(any interface{}, typeName, source string, changeCount
 
 func (this *Cache) createAddNotification(any interface{}) (*types.NotificationSet, error) {
 	defer func() { this.sequence++ }()
-	return CreateAddNotification(any, this.typeName, this.source, 1, this.sequence)
+	return CreateAddNotification(any, this.multicastGroup, this.protoType, this.source, 1, this.sequence)
 }
 
-func CreateReplaceNotification(old, new interface{}, typeName, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
-	notificationSet := CreateNotificationSet(types.NotificationType_Replace, typeName, source, 1, sequence)
+func CreateReplaceNotification(old, new interface{}, multicastGroup, protoType, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
+	notificationSet := CreateNotificationSet(types.NotificationType_Replace, multicastGroup, protoType, source, 1, sequence)
 	oldObj := object.NewEncode([]byte{}, 0)
 	err := oldObj.Add(old)
 	if err != nil {
@@ -65,11 +66,11 @@ func CreateReplaceNotification(old, new interface{}, typeName, source string, ch
 
 func (this *Cache) createReplaceNotification(old, new interface{}) (*types.NotificationSet, error) {
 	defer func() { this.sequence++ }()
-	return CreateReplaceNotification(old, new, this.typeName, this.source, 1, this.sequence)
+	return CreateReplaceNotification(old, new, this.multicastGroup, this.protoType, this.source, 1, this.sequence)
 }
 
-func CreateDeleteNotification(any interface{}, typeName, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
-	notificationSet := CreateNotificationSet(types.NotificationType_Delete, typeName, source, 1, sequence)
+func CreateDeleteNotification(any interface{}, multicastGroup, protoType, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
+	notificationSet := CreateNotificationSet(types.NotificationType_Delete, multicastGroup, protoType, source, 1, sequence)
 	obj := object.NewEncode([]byte{}, 0)
 	err := obj.Add(any)
 	if err != nil {
@@ -83,11 +84,11 @@ func CreateDeleteNotification(any interface{}, typeName, source string, changeCo
 
 func (this *Cache) createDeleteNotification(any interface{}) (*types.NotificationSet, error) {
 	defer func() { this.sequence++ }()
-	return CreateDeleteNotification(any, this.typeName, this.source, 1, this.sequence)
+	return CreateDeleteNotification(any, this.multicastGroup, this.protoType, this.source, 1, this.sequence)
 }
 
-func CreateUpdateNotification(changes []*updating.Change, typeName, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
-	notificationSet := CreateNotificationSet(types.NotificationType_Update, typeName, source, changeCount, sequence)
+func CreateUpdateNotification(changes []*updating.Change, multicastGroup, protoType, source string, changeCount int, sequence uint32) (*types.NotificationSet, error) {
+	notificationSet := CreateNotificationSet(types.NotificationType_Update, multicastGroup, protoType, source, changeCount, sequence)
 	for i, change := range changes {
 		n := &types.Notification{}
 		n.PropertyId = change.PropertyId()
@@ -114,7 +115,7 @@ func CreateUpdateNotification(changes []*updating.Change, typeName, source strin
 
 func (this *Cache) createUpdateNotification(changes []*updating.Change) (*types.NotificationSet, error) {
 	defer func() { this.sequence++ }()
-	return CreateUpdateNotification(changes, this.typeName, this.source, len(changes), this.sequence)
+	return CreateUpdateNotification(changes, this.multicastGroup, this.protoType, this.source, len(changes), this.sequence)
 }
 
 func ItemOf(n *types.NotificationSet, i common.IIntrospector) (interface{}, error) {
@@ -122,15 +123,15 @@ func ItemOf(n *types.NotificationSet, i common.IIntrospector) (interface{}, erro
 	case types.NotificationType_Replace:
 		fallthrough
 	case types.NotificationType_Add:
-		obj := object.NewDecode(n.NotificationList[0].NewValue, 0, n.TypeName, i.Registry())
+		obj := object.NewDecode(n.NotificationList[0].NewValue, 0, n.ProtoType, i.Registry())
 		v, e := obj.Get()
 		return v, e
 	case types.NotificationType_Delete:
-		obj := object.NewDecode(n.NotificationList[0].OldValue, 0, n.TypeName, i.Registry())
+		obj := object.NewDecode(n.NotificationList[0].OldValue, 0, n.ProtoType, i.Registry())
 		v, e := obj.Get()
 		return v, e
 	case types.NotificationType_Update:
-		info, err := i.Registry().Info(n.TypeName)
+		info, err := i.Registry().Info(n.ProtoType)
 		if err != nil {
 			return nil, err
 		}
